Pulls detail::AABB, detail::underlying_type out to headers under detail directory
